Separar las interfaz de productos de inventory_management
